Update obsoletes and approve API
