+++ src/NServiceBus.SqlServer/SqlServerTransportSettingsExtensions.cs
@@ -226,20 +226,6 @@
             return transportExtensions;
         }
 
-        /// <summary>
-        /// Enables multi-instance mode.
-        /// </summary>
-        /// <param name="transportExtensions">The <see cref="TransportExtensions{T}" /> to extend.</param>
-        /// <param name="sqlConnectionFactory">Function that returns opened sql connection based on destination queue name.</param>
-        [ObsoleteEx(
-            RemoveInVersion = "5.0",
-            TreatAsErrorFromVersion = "4.0",
-            Message = "Multi-instance mode has been deprecated. Use Transport Bridge and/or multi-catalog addressing instead.")]
-        public static TransportExtensions<SqlServerTransport> EnableLegacyMultiInstanceMode(this TransportExtensions<SqlServerTransport> transportExtensions, Func<string, Task<SqlConnection>> sqlConnectionFactory)
-        {
-            throw new NotImplementedException();
-        }
-
         static ILog Logger = LogManager.GetLogger<SqlServerTransport>();
     }
 }
